+++ Presentation/OpenTgResearcherConsole/Helpers/TgMenuHelperClientConDownload.cs
@@ -253,7 +253,7 @@ internal partial class TgMenuHelper
 
     private async Task DownloadSettingsSaveAsync(TgDownloadSettingsViewModel tgDownloadSettings)
     {
-        await tgDownloadSettings.UpdateSourceWithSettingsAsync();
+        await tgDownloadSettings.SourceVm.SaveAsync();
         await BusinessLogicManager.ConnectClient.UpdateStateSourceAsync(tgDownloadSettings.SourceVm.Dto.Id, tgDownloadSettings.SourceVm.Dto.FirstId, tgDownloadSettings.SourceVm.Dto.Count,
             TgLocale.SettingsChat);
     }