+++ Docs/CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.680] - 2025-07-18
+## [0.6.690] - 2025-07-18
 ### Fixed
 - Fixed online chat update in OpenTgResearcherDesktop
 - Fixed chats page in OpenTgResearcherDesktop
 - Fixed client connection page in OpenTgResearcherDesktop
 - Fixed settings page in OpenTgResearcherDesktop
+- Fixed client connection
 ### Changed
 - Updated message table in TgStorage
 - Updated the main page in OpenTgResearcherDesktop